refactor entity set update (patch and put) operation handler classes to use same base class
