+++ src/Microsoft.OpenApi.OData.Reader/Operation/EntityPatchOperationHandler.cs
@@ -3,14 +3,7 @@
 //  Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // ------------------------------------------------------------
 
-using System.Collections.Generic;
-using System.Linq;
-using Microsoft.OData.Edm;
 using Microsoft.OpenApi.Models;
-using Microsoft.OpenApi.OData.Common;
-using Microsoft.OpenApi.OData.Edm;
-using Microsoft.OpenApi.OData.Generator;
-using Microsoft.OpenApi.OData.Vocabulary.Capabilities;
 
 namespace Microsoft.OpenApi.OData.Operation
 {
@@ -19,105 +12,9 @@ namespace Microsoft.OpenApi.OData.Operation
     /// The Path Item Object for the entity set contains the keyword patch with an Operation Object as value
     /// that describes the capabilities for updating the entity.
     /// </summary>
-    internal class EntityPatchOperationHandler : EntitySetOperationHandler
+    internal class EntityPatchOperationHandler : EntityUpdateOperationHandler
     {
         /// <inheritdoc/>
         public override OperationType OperationType => OperationType.Patch;
-
-        /// <inheritdoc/>
-        protected override void SetBasicInfo(OpenApiOperation operation)
-        {
-            // Summary
-            operation.Summary = "Update entity in " + EntitySet.Name;
-
-            IEdmEntityType entityType = EntitySet.EntityType();
-
-            // Description
-            operation.Description = Context.Model.GetDescriptionAnnotation(entityType);
-
-            // OperationId
-            if (Context.Settings.EnableOperationId)
-            {
-                string typeName = entityType.Name;
-                operation.OperationId = EntitySet.Name + "." + typeName + ".Update" + Utils.UpperFirstChar(typeName);
-            }
-        }
-
-        /// <inheritdoc/>
-        protected override void SetRequestBody(OpenApiOperation operation)
-        {
-            OpenApiSchema schema = null;
-
-            if (Context.Settings.EnableDerivedTypesReferencesForRequestBody)
-            {
-                schema = EdmModelHelper.GetDerivedTypesReferenceSchema(EntitySet.EntityType(), Context.Model);
-            }
-
-            if (schema == null)
-            {
-                schema = new OpenApiSchema
-                {
-                    Reference = new OpenApiReference
-                    {
-                        Type = ReferenceType.Schema,
-                        Id = EntitySet.EntityType().FullName()
-                    }
-                };
-            }
-
-            operation.RequestBody = new OpenApiRequestBody
-            {
-                Required = true,
-                Description = "New property values",
-                Content = new Dictionary<string, OpenApiMediaType>
-                {
-                    {
-                        Constants.ApplicationJsonMediaType, new OpenApiMediaType
-                        {
-                            Schema = schema
-                        }
-                    }
-                }
-            };
-
-            base.SetRequestBody(operation);
-        }
-
-        /// <inheritdoc/>
-        protected override void SetResponses(OpenApiOperation operation)
-        {
-            operation.AddErrorResponses(Context.Settings, true);
-            base.SetResponses(operation);
-        }
-
-        protected override void SetSecurity(OpenApiOperation operation)
-        {
-            UpdateRestrictionsType update = Context.Model.GetRecord<UpdateRestrictionsType>(EntitySet, CapabilitiesConstants.UpdateRestrictions);
-            if (update == null || update.Permissions == null)
-            {
-                return;
-            }
-
-            operation.Security = Context.CreateSecurityRequirements(update.Permissions).ToList();
-        }
-
-        protected override void AppendCustomParameters(OpenApiOperation operation)
-        {
-            UpdateRestrictionsType update = Context.Model.GetRecord<UpdateRestrictionsType>(EntitySet, CapabilitiesConstants.UpdateRestrictions);
-            if (update == null)
-            {
-                return;
-            }
-
-            if (update.CustomHeaders != null)
-            {
-                AppendCustomParameters(operation, update.CustomHeaders, ParameterLocation.Header);
-            }
-
-            if (update.CustomQueryOptions != null)
-            {
-                AppendCustomParameters(operation, update.CustomQueryOptions, ParameterLocation.Query);
-            }
-        }
     }
 }
